fastapi_app/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
openai==1.12.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pytest==8.0.1
httpx==0.26.0
sqlalchemy==2.0.27
aio-pika==9.4.1
python-multipart
aiokafka==0.10.0
websockets==12.0
confluent-kafka==2.3.0 